fix: criação do service user_role
